+++ deepdow/data/synthetic.py
@@ -1,15 +1,10 @@
-"""Collection of functions generating synthetic datasets.
-
-All these functions are outputting a 3D ``np.ndarray`` of shape
- `(n_timesteps, n_channels, n_assets)`."""
+"""Collection of functions generating synthetic datasets."""
 
 import numpy as np
 
 
 def sin_single(n_timesteps, amplitude=1, freq=0.25, phase=0):
     """Generate sine waves.
-
-    Of form :math:`amplitude * /sin(2\pi(x) * freq + phase)`.
 
     Parameters
     ----------
@@ -20,7 +15,7 @@ def sin_single(n_timesteps, amplitude=1, freq=0.25, phase=0):
         The peak value.
 
     freq : float
-        Frequency - number of oscilations per timestep.
+        Frequency - number of oscillations per timestep.
 
     phase : float
         Offset.
@@ -28,7 +23,7 @@ def sin_single(n_timesteps, amplitude=1, freq=0.25, phase=0):
     Returns
     -------
     y : np.ndarray
-        1D array of shape `(n_timesteps,)`
+        1D array of shape `(n_timesteps,)`.
 
     """
     x = np.arange(n_timesteps)
